requirements.txt
requests
sentence-transformers
scikit-learn
torch
ultralytics
opencv-python
openai-whisper
TTS
sounddevice
google-generativeai